api/requirements.txt
# Core web framework - Vercel compatible versions
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI and ML
openai>=1.12.0
numpy>=1.26.0

# Environment and utilities
python-dotenv==1.0.0
mmh3==3.0.0

# HTTP and networking
requests==2.31.0
httpx>=0.23,<0.24

# Templates and forms
jinja2==3.1.2
python-multipart==0.0.6

# Authentication
supabase==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2

# Data validation
pydantic[email]>=2.11.7,<3.0.0

# Additional dependencies 